Improve RGBA colorization performance.
got rid of Qt's QGraphicsColorizeEffect as it's horribly memory inefficent when zoomed in on an image (because of transformation it's probably making a huge image and colorizing it).

The way it works now is by having a QPixmap as a solid color and painting that over the original image with a multiply blend mode. (ALL of this doesn't happen if color strength is 0, for other use cases to be efficent.)