Changed the way Events and Actions are implemented in the editor. check the description..
Instead of having a seperate dialog class for every event and action
now there are only 2, one for the events and the other for the actions. each have a stacked widget that sets it's page index based on the EventType enum that is passed to the dialog.

much more managable now, and less tedious with the whole passing parameters around... hope that makes sense.